+++ Projekt/Interfaces/StockMarket.java
@@ -5,17 +5,23 @@ import Exceptions.StockMarketException;
 import java.util.Set;
 
 public interface StockMarket {
-    long getPrice(String stockName) throws StockMarketException;
+    int getPrice(String stockName) throws StockMarketException;
 
-    long purchase(String stockName) throws StockMarketException;
+    int purchase(String stockName) throws StockMarketException;
 
-    long purchase(String stockName, long amount) throws StockMarketException;
+    int purchase(String stockName, int amount) throws StockMarketException;
 
-    long sell(String stockName) throws StockMarketException;
+    int sell(String stockName) throws StockMarketException;
 
-    long sell(String stockName, long amount) throws StockMarketException;
+    int sell(String stockName, int amount) throws StockMarketException;
 
 //    void globalUpdate(double percentage);
+
+    /**
+     * Add or update a stock to the market
+     * @param stockName Name of the stock to add or update
+     * @param stock Instance of the stock to add or update
+     */
 
     void updateStock(String stockName, Stock stock);
